Create a call set for thumbnails requested
Retracer now knows what thumbnails are needed. This information needs
to be communicated to the glretrace command.

Construct a string that contains the call numbers for all missing thumbnails.

The constructed string is then passed to glretrace with the "-S" argument.

TODO: we want to be able to detect contiguous sequences of calls and emit
them as a range. However, I wasn't able to get this working correctly and
will defer this to a future patch.